+++ compatibility_spark-3/src/test/scala/za/co/absa/hyperdrive/compatibility/impl/writer/delta/scd2/TestDeltaCDCToSCD2Writer.scala
@@ -15,31 +15,12 @@
 
 package za.co.absa.hyperdrive.compatibility.impl.writer.delta.scd2
 
-import org.apache.spark.sql.execution.streaming.MemoryStream
 import org.apache.spark.sql.streaming.Trigger
 import org.scalatest.mockito.MockitoSugar
-import org.scalatest.{BeforeAndAfterEach, FlatSpec, Matchers}
-import za.co.absa.commons.io.TempDirectory
-import za.co.absa.hyperdrive.compatibility.impl.writer.delta.{CDCEvent, DeltaEvent}
-import za.co.absa.hyperdrive.shared.utils.SparkTestBase
+import org.scalatest.{FlatSpec, Matchers}
+import za.co.absa.hyperdrive.compatibility.impl.writer.delta.{CDCEvent, DeltaEvent, DeltaTestBase}
 
-import scala.reflect.io.Path
-
-class TestDeltaCDCToSCD2Writer extends FlatSpec with MockitoSugar with Matchers with BeforeAndAfterEach with SparkTestBase {
-  private val baseDir: TempDirectory = TempDirectory("TestDeltaCDCToSCD2Writer").deleteOnExit()
-  private val destinationPath = s"${baseDir.path.toAbsolutePath.toString}/destination"
-  private val checkpointPath = s"${baseDir.path.toAbsolutePath.toString}/checkpoint"
-
-  import spark.implicits._
-
-  private val memoryStream = MemoryStream[CDCEvent](1, spark.sqlContext)
-
-  override def beforeEach(): Unit = {
-    Path(destinationPath).deleteRecursively()
-    Path(checkpointPath).deleteRecursively()
-    memoryStream.reset()
-  }
-
+class TestDeltaCDCToSCD2Writer extends FlatSpec with MockitoSugar with Matchers with DeltaTestBase {
   behavior of "DeltaCDCToSCD2Writer"
 
   it should "merge unique by timestamp scd2 rows into empty delta table" in {
@@ -122,9 +103,4 @@ class TestDeltaCDCToSCD2Writer extends FlatSpec with MockitoSugar with Matchers 
     precombineColumnsCustomOrder = Map("eventType" -> Seq("PT", "FI", "RR", "UB", "UP", "DL", "FD")),
     extraConfOptions = Map.empty[String, String]
   )
-
-  private def getResult: Seq[DeltaEvent] = {
-    import spark.implicits._
-    spark.read.format("delta").load(destinationPath).as[DeltaEvent].collect().toSeq
-  }
 }
